V2

+++ lib/presentation/pages/workout/workout_page.dart
@@ -1,3 +1,4 @@
+import 'package:animations/animations.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
@@ -46,9 +47,9 @@ class WorkoutPage extends StatelessWidget {
           actions: [
             RemoveButton(
               onPressed: () {
-                showDialog(
+                showModal(
                   context: context,
-                  builder: (BuildContext context) {
+                  builder: (_) {
                     return DeleteConfirmDialog(exerciseIds: [exercise.id]);
                   },
                 );
